requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
Django==4.2.2
django-celery-results==2.5.1
django-environ==0.9.0
django-filter==23.2
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
idna==3.4
importlib-metadata==6.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
Markdown==3.4.3
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.4
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
